requirements.txt
loguru>=0.7.2
PySide6>=6.7.2
pyusb>=1.2.1
PyYAML>=6.0.2
libusb-package>=1.0.26.1; platform_system == "Windows"
